Centralize common database helpers
