+++ cogs/rules/service/rules_services.py
@@ -2,40 +2,10 @@
 import logging
 from typing import Optional, Dict
 from utils.database import database
+from services.base import get_or_create_server_record, store_persistent_message, get_persistent_message, delete_persistent_message
 
 logger = logging.getLogger("rules_service")
 
-# ------------------------------------------------
-# 1) Obtenir ou créer un enregistrement de serveur
-# ------------------------------------------------
-async def get_or_create_server_record(discord_guild_id: int, guild_name: str = "Inconnu") -> Optional[int]:
-    """
-    Récupère ou crée l'ID interne (PK) dans la table serveur_id 
-    pour un guild_id (Discord) donné.
-    """
-    try:
-        select_query = """
-            SELECT id
-              FROM serveur_id
-             WHERE guild_id = $1
-        """
-        record = await database.fetchrow(select_query, discord_guild_id)
-        if record:
-            return record["id"]
-
-        insert_query = """
-            INSERT INTO serveur_id (guild_id, serveur)
-            VALUES ($1, $2)
-            RETURNING id;
-        """
-        new_id = await database.fetchval(insert_query, discord_guild_id, guild_name)
-        logger.info(f"[get_or_create_server_record] Serveur créé: guild_id={discord_guild_id}, id={new_id}")
-        return new_id
-    except Exception as e:
-        logger.error(f"[get_or_create_server_record] Erreur : {e}")
-        return None
-
-# ------------------------------------------------
 # 2) Gérer le salon "rules"
 # ------------------------------------------------
 async def get_rules_channel_id(discord_guild_id: int, guild_name: str = "Inconnu") -> Optional[int]:
@@ -72,62 +42,15 @@ async def store_rules_message(discord_guild_id: int,
                              guild_name: str,
                              channel_id: int,
                              message_id: int) -> bool:
-    """
-    Stocke (ou met à jour) un message persistant de type 'rules_embed' 
-    dans la table persistent_messages.
-    """
-    try:
-        server_db_id = await get_or_create_server_record(discord_guild_id, guild_name)
-        if not server_db_id:
-            return False
+    """Enregistre le message persistant de type ``rules_embed``."""
+    return await store_persistent_message(
+        discord_guild_id,
+        channel_id,
+        message_id,
+        "rules_embed",
+        guild_name,
+    )
 
-        query = """
-            INSERT INTO persistent_messages (guild_id, channel_id, message_id, message_type)
-            VALUES ($1, $2, $3, 'rules_embed')
-            ON CONFLICT (guild_id, message_type) DO UPDATE
-               SET channel_id = EXCLUDED.channel_id,
-                   message_id = EXCLUDED.message_id,
-                   created_at = NOW();
-        """
-        await database.execute(query, server_db_id, channel_id, message_id)
-        logger.info(f"[store_rules_message] rules_embed stocké: server_id={server_db_id}, msg={message_id}")
-        return True
-    except Exception as e:
-        logger.error(f"[store_rules_message] Erreur: {e}")
-        return False
-
-async def get_persistent_message(discord_guild_id: int,
-                                message_type: str,
-                                guild_name: str = "Inconnu") -> Optional[Dict[str, int]]:
-    """
-    Récupère les informations d'un message persistant spécifique.
-    """
-    try:
-        server_db_id = await get_or_create_server_record(discord_guild_id, guild_name)
-        if not server_db_id:
-            return None
-
-        query = """
-            SELECT channel_id, message_id 
-              FROM persistent_messages 
-             WHERE guild_id = $1
-               AND message_type = $2
-        """
-        record = await database.fetchrow(query, server_db_id, message_type)
-        if record:
-            logger.debug(
-                f"[get_persistent_message] Message récupéré: server_db_id={server_db_id}, "
-                f"type={message_type}, channel_id={record['channel_id']}, message_id={record['message_id']}"
-            )
-            return {'channel_id': record['channel_id'], 'message_id': record['message_id']}
-        else:
-            logger.warning(
-                f"[get_persistent_message] Aucun message trouvé pour server_db_id={server_db_id}, type={message_type}."
-            )
-            return None
-    except Exception as e:
-        logger.error(f"[get_persistent_message] Erreur : {e}")
-        return None
 
 # ------------------------------------------------
 # 4) Enregistrement des utilisateurs
@@ -168,30 +91,3 @@ async def has_accepted_rules(discord_id: int) -> bool:
         return False
 
 
-async def delete_persistent_message(discord_guild_id: int,
-                                   message_type: str,
-                                   guild_name: str = "Inconnu") -> bool:
-    """
-    Supprime les informations d'un message persistant de la base de données.
-
-    :param discord_guild_id: ID Discord brut de la guilde.
-    :param message_type: Type du message.
-    :param guild_name: Nom du serveur, par défaut "Inconnu".
-    :return: True si réussi, False sinon.
-    """
-    try:
-        server_db_id = await get_or_create_server_record(discord_guild_id, guild_name)
-        if not server_db_id:
-            return False
-
-        query = """
-            DELETE FROM persistent_messages 
-             WHERE guild_id = $1
-               AND message_type = $2;
-        """
-        await database.execute(query, server_db_id, message_type)
-        logger.info(f"Message persistant supprimé: server_id={server_db_id}, type={message_type}")
-        return True
-    except Exception as e:
-        logger.error(f"Erreur lors de la suppression du message persistant: {e}")
-        return False
